Split tests by concept
