+++ Packages/EurofurenceKit/Tests/EurofurenceKitTests/Model Ingestion/IngestingFullRemoteModel_Errors.swift
@@ -1,29 +1,9 @@
-@testable import EurofurenceKit
+import EurofurenceKit
 import Logging
 import XCTest
 
-class EurofurenceKitTests: XCTestCase {
-    
-    func testIngestingRemoteResponse() async throws {
-        let logger = Logger(label: "Test")
-        let network = FakeNetwork()
-        let configuration = EurofurenceModel.Configuration(
-            environment: .memory,
-            network: network,
-            logger: logger,
-            conventionIdentifier: ConventionIdentifier("EF26")
-        )
-        
-        let model = EurofurenceModel(configuration: configuration)
-        
-        let sampleResponse = EF26FullSyncResponseFile()
-        let syncURL = try XCTUnwrap(URL(string: "https://app.eurofurence.org/EF26/api/Sync"))
-        network.stub(url: syncURL, with: try sampleResponse.loadFileContents())
-        try await model.updateLocalStore()
-        
-        try sampleResponse.assertAgainstEntities(in: model.viewContext)
-    }
-    
+class IngestingFullRemoteModel_Errors: XCTestCase {
+
     func testIngestingRemoteResponse_RemoteFetchFailure() async throws {
         let logger = Logger(label: "Test")
         let network = FakeNetwork()
@@ -66,5 +46,5 @@ class EurofurenceKitTests: XCTestCase {
             { try await model.updateLocalStore() }
         )
     }
-    
+
 }
